Update EditCommand and UniquePersonList to correctly detect duplicate persons
Currently, an error for a duplicate person is
only thrown if editedPerson:
- is not the same as personToEdit
and
- is the same as a person in addressbook

However, since isSamePerson() returns true if
2 Person objects:
- share the same student ID
OR
- the same email
it is possible for editedPerson to be
the same person as both personToEdit and another
person in addressbook, which will not throw
an error.

To fix this, change UniquePersonList.setPerson to
check for duplicates by checking if a filtered
stream of its internal list without the target
person contains any person matching editedPerson.

Change EditCommand.execute to use the same check
by calling setPerson in a try-catch block.

Signed-off-by: yoyongxuan <66002397+yoyongxuan@users.noreply.github.com>