Update isValidScore to handle large integers
isValidScore currently returns true when an
input string representing an integer larger
than the max integer value is given.

This causes a NumberFormatException to occur
when the program tries to parse this string
into an integer, assuming it represents a valid
integer

Update isValidScore to return False when an
input string representing an integer larger
than the max integer value is given.

Signed-off-by: yoyongxuan <66002397+yoyongxuan@users.noreply.github.com>